add test for RecoQuery
