+++ lib/segment/src/vector_storage/query_scorer/mod.rs
@@ -2,6 +2,7 @@ use crate::data_types::vectors::VectorElementType;
 use crate::types::{PointOffsetType, ScoreType};
 
 pub mod metric_query_scorer;
+pub(crate) mod reco_query;
 pub mod reco_query_scorer;
 
 pub trait QueryScorer {
